add support for sessions
when a user logs in they gain a new "session".
A session is a token associated with their user.
The session is stored in a browser cookie.
It can be read by request handlers in the future to confirm a user's
identity.
Currently sessions expire after a week, ideally this would be configurable.